% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add logo}
\usage{
add_logo(
  path_viz,
  path_logo,
  idx_x,
  idx_y,
  logo_scale = 0.1,
  adjust_x = TRUE,
  adjust_y = TRUE,
  path_suffix = "_w_logo",
  delete = TRUE
)
}
\description{
Add logo to a plot saved as a file
}
