% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_sliced}
\alias{palette_sliced}
\title{Color palette for SLICED}
\usage{
palette_sliced(n)
}
\arguments{
\item{n}{Number of different colors in the palette, minimum 3, maximum is 11 (like \code{\link[RColorBrewer]{brewer.pal}} with "PRGn")}
}
\description{
Color palette for SLICED
}
